+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix handling of sub folders when copying files (in particular the ATEM ISO structure). Was flattening the folder structures.
+
+### Added
+- Add --no-log-identical-operations to quieten down logs a bit
+- Log failures at end of sync
 
 ## [0.7.0] - 2024-09-18
 ### Added